# Optimize cStringArrayToSlice python check helper