<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
-->
### What does this PR do?

* Use string interning for cgo strings
* Eliminate unnecessary cgo call to `C.getStringAddr`
* Pre-size string slice returned by `cStringArrayToSlice`
* Add a benchmark to estimate the impact

```
goos: linux
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/collector/python
                        │  before.txt  │              after.txt              │
                        │    sec/op    │   sec/op     vs base                │
_cStringArrayToSlice-10   1189.0n ± 5%   221.9n ± 2%  -81.34% (p=0.000 n=10)

                        │ before.txt │             after.txt              │
                        │    B/op    │    B/op     vs base                │
_cStringArrayToSlice-10   624.0 ± 0%   160.0 ± 0%  -74.36% (p=0.000 n=10)

                        │ before.txt  │             after.txt              │
                        │  allocs/op  │ allocs/op   vs base                │
_cStringArrayToSlice-10   15.000 ± 0%   1.000 ± 0%  -93.33% (p=0.000 n=10)
```

### Motivation

My estimate is that this change will reduce the agent CPU usage on staging by up to 6% due to:

* ~3% CPU reduction by direct optimizations of cStringArrayToSlice
* ~3% CPU reduction by up to 35% reduction of GC work

This estimate is based on the data from this [notebook](https://ddstaging.datadoghq.com/notebook/7645357/prof-9301-optimize-allocation-of-cgo-strings-from-python-checks?viz=flame_graph). The GC cost reduction calculation can be found in this [spreadsheet](https://docs.google.com/spreadsheets/d/1-OS4-Z2-vW_UB0QeB8XnILhNYigNt-gI4hw7L8tSdRI/edit#gid=0).

The real impact will probably be a little lower, but hopefully not by much. I'm hoping to see 4-5%.

### Additional Notes

I was hoping the `basic_py_check` introduced in https://github.com/DataDog/single-machine-performance/pull/1671 would detect the improvements made in this PR, but looking at [the profile](https://app.datadoghq.com/profiling/explorer?query=experiment%3Abasic_py_check%20service%3Adatadog-agent%20git.commit.sha%3A045ed1b99573ecd7cd5a054bc0c4688dc99b2575%20&fromUser=true&my_code=disabled&viz=flame_graph&start=1709987282572&end=1709988514895&paused=true), it looks like it's creating almost no python check load, so it's not detecting anything. But it's likely that I misunderstand the purpose of this check 😅 

### Possible Drawbacks / Trade-offs

Bugs in the code could lead to crashes or memory corruption.

### Describe how to test/QA your changes

So far I've only estimated the impact. I guess the next step could be to deploy this to staging?

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->