@iksaif 2024-03-11
This looks very nice, but we should probably also have a `SubmitMetrics` that sends multiple metrics sharing the same tags and host :)