@felixge 2024-03-11
> This looks very nice, but we should probably also have a `SubmitMetrics` that sends multiple metrics sharing the same tags and host :)

Oh, that sounds like a very good idea as well. I'll admit to not having taken a close look at that part of the call tree yet 😅. Would this require patching a lot of python code?